+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Update to Default Library `Index` & `Bib` [BFP-381]
 - Change the XML structure of `ComplexSubjects`
+- Add Complex Lookup for LCDGT - `Creator Characteristic`
 
 
 ## [1.2.6] - 2025-04-08